Switch to using Swedish layout in OS
Instead of changing layout to US in the OS (becomes annoying having to
switch back and forth when using the laptop keyboard), just "reverse the
mapping", i.e., send the key codes that would be used on a swedish
keyboard, but from where the keys are placed in a US layout.
Since `, ~, and ^ are "dead keys" on a swedish keyboard, send them with
a space after to "revive" them.